requirements.txt
contextvars==2.4
fastapi[all]>=0.54.1
python-jose[rsa]>=3.1.0
requests>=2.25.1
starlette>=0.13.2
